requirements.txt
langchain-openai==0.1.0
pdfplumber==0.5.28
pillow==8.4.0
pymupdf==1.18.22
pytesseract==0.3.9
pandas==1.3.3
streamlit==1.4.0
